Add comprehensive tests for background cache warming
- 9 new unit tests for BackgroundTaskManager
- Test singleton pattern, config registration, lifecycle
- Test task start/stop/restart scenarios
- Test error handling and graceful failures
- Support float interval_hours for flexible testing
- All 42 tests passing (33 original + 9 new)
- Verified background warming triggers on catalog access

+++ app/services/background.py
@@ -67,7 +67,7 @@ class BackgroundTaskManager:
         
         logger.info("Cache warming cycle complete")
     
-    async def background_loop(self, interval_hours: int = 3):
+    async def background_loop(self, interval_hours: float = 3):
         """
         Background loop that warms caches periodically
         
@@ -90,7 +90,7 @@ class BackgroundTaskManager:
                 logger.error(f"Error in background loop: {e}", exc_info=True)
                 # Continue running despite errors
     
-    def start(self, interval_hours: int = 3):
+    def start(self, interval_hours: float = 3):
         """Start the background task"""
         if self.task is None or self.task.done():
             self.task = asyncio.create_task(self.background_loop(interval_hours))
